refactor(stats): create charts and tables directories
