+++ src/components/stats/charts/SurvivalRateChart.tsx
